refactor: Update project page to conditionally disable quotation button
The code changes in this commit update the project page to conditionally disable the "ใบเสนอราคา" (quotation) button based on the status of the BOQ (Bill of Quantities) for the project. If the BOQ status is "approved", the button is enabled and links to the quotation page. Otherwise, the button is disabled.

This refactor improves the user experience by providing visual feedback and preventing users from accessing the quotation page when the BOQ is not approved.